requirements/test.txt
-r develop.txt

pytest
pytest-django
coverage